fix(runtime): session channel send fallibility, IO flush-before-shutdown, NATS sync close, timer wheel modulo
Four independent correctness fixes across runtime subsystems:

1. TrackedPermit::send() now returns Result<CommittedProof, SendError>
   instead of infallibly consuming the permit. If the receiver was dropped
   between reserve() and send(), the obligation is properly aborted rather
   than committed for a delivery that never happened. Tests updated to
   unwrap the new Result.

2. Bidirectional IO copy (copy.rs) now flushes the write side before
   initiating shutdown, in both A→B and B→A transfer directions. Without
   this, buffered data could be lost when the write side has pending bytes
   at the time poll_shutdown is called.

3. HTTP/1.1 server (server.rs) now gracefully shuts down the connection
   via AsyncWriteExt::shutdown() after the request loop completes, ensuring
   TLS close_notify is sent and the peer observes clean connection closure.

4. NatsClient::close() converted from async to sync (the body was already
   synchronous, evidenced by the clippy::unused_async allow). Now also
   calls stream.shutdown(Both) to close the TCP socket immediately rather
   than relying on drop ordering.

5. Timer wheel modulo fix (intrusive_wheel.rs + wheel.rs): slot index
   computation now applies modulo before casting to usize, preventing
   truncation on 32-bit targets where large tick values would wrap
   incorrectly. Changed from `(tick as usize) % SLOTS` to
   `(tick % (SLOTS as u128/u64)) as usize`.

6. src/types/cancel.rs: rustfmt on severity test assertions.

+++ src/time/intrusive_wheel.rs
@@ -395,7 +395,7 @@ impl<const SLOTS: usize> TimerWheel<SLOTS> {
         let elapsed = deadline.saturating_duration_since(self.base_time);
         let ticks = elapsed.as_nanos() / self.resolution.as_nanos().max(1);
         let safe_ticks = ticks.max(u128::from(self.current_tick));
-        (safe_ticks as usize) % SLOTS
+        (safe_ticks % (SLOTS as u128)) as usize
     }
 
     /// Inserts a timer node with the given deadline.
@@ -485,9 +485,9 @@ impl<const SLOTS: usize> TimerWheel<SLOTS> {
                 self.count = self.count.saturating_sub(wakers.len());
                 all_wakers.extend(wakers);
             }
-            self.current = (target_tick_u64 as usize) % SLOTS;
+            self.current = (target_tick_u64 % (SLOTS as u64)) as usize;
         } else {
-            let target_slot = (target_tick_u64 as usize) % SLOTS;
+            let target_slot = (target_tick_u64 % (SLOTS as u64)) as usize;
 
             // Process slots until we reach target (handling wrap-around)
             while self.current != target_slot {
